agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.136.0
pydantic==2.12.5
cloudpickle==3.1.2
google-genai==1.62.0
python-dotenv==1.0.1
requests==2.32.4
python-multipart==0.0.20
